+++ CHANGELOG.md
@@ -703,7 +703,7 @@ Huge Thank You for this release to our contributors:
 * Upgrade ingress controller to v0.9.0 [#2292](https://github.com/kubernetes/minikube/pull/2292)
 * Upgrade docker machine to g49dfaa70 [#2299](https://github.com/kubernetes/minikube/pull/2299)
 * Added ingress integration tests [#2254](https://github.com/kubernetes/minikube/pull/2254)
-* Converted image registries to k8s.gcr.io [#2356](https://github.com/kubernetes/minikube/pull/2356)
+* Converted image registries to gcr.azk8s.cn/google_containers [#2356](https://github.com/kubernetes/minikube/pull/2356)
 * Added cache list command [#2272](https://github.com/kubernetes/minikube/pull/2272)
 * Upgrade to Kubernetes 1.9 [#2343](https://github.com/kubernetes/minikube/pull/2343)
 * [hyperkit] Support NFS Sharing [#2337](https://github.com/kubernetes/minikube/pull/2337)